# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Role and Permission Management API",
        "description": "API для управления ролями и привилегиями",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users/{userId}/roles": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Назначить роль пользователю",
                "operationId": "3efb0f8b78abf54f96b21d3546512f45",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "editor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Роль успешно назначена"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/api/roles/{roleId}/permissions": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Назначить привилегии роли",
                "operationId": "f9413000bd2e6aca7e16454706685e05",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "description": "ID роли",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "create-post"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Привилегии успешно назначены"
                    },
                    "404": {
                        "description": "Роль не найдена"
                    }
                }
            }
        },
        "/api/products": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Создать продукт",
                "operationId": "9934bcf87d6613c14890dbdd587d311a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Продукт 1"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Продукт успешно создан"
                    }
                }
            }
        },
        "/api/users/{userId}/remove-buyer": {
            "post": {
                "tags": [
                    "TeamLead"
                ],
                "summary": "Удалить роль баера",
                "operationId": "3a20e1498d03a793396e12d66010e636",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Роль баера успешно удалена"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Admin",
            "description": "Управление ролями и привилегиями"
        },
        {
            "name": "TeamLead",
            "description": "Управление баерами"
        },
        {
            "name": "Product",
            "description": "Product"
        }
    ]
}